requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-decouple==3.8
python-dateutil==2.8.2
stripe==7.0.0
Pillow==10.1.0
